refactor(data): 불필요한 패키지 제거
