+++ app/src/main/java/woowacourse/paint/presentation/bindingAdapter/PaintBoardViewBindingAdapter.kt
@@ -3,9 +3,9 @@ package woowacourse.paint.presentation.bindingAdapter
 import androidx.core.content.ContextCompat
 import androidx.databinding.BindingAdapter
 import woowacourse.paint.R
-import woowacourse.paint.data.model.Brush
-import woowacourse.paint.data.model.Color
 import woowacourse.paint.presentation.curstomView.PaintBoardView
+import woowacourse.paint.presentation.ui.model.Brush
+import woowacourse.paint.presentation.ui.model.Color
 
 @BindingAdapter("app:onTouch")
 fun PaintBoardView.setOnTouchListener(listener: PaintBoardView.TouchEventListener) {
