+++ app/src/main/java/woowacourse/paint/presentation/ui/model/Color.kt
@@ -1,4 +1,4 @@
-package woowacourse.paint.data.model
+package woowacourse.paint.presentation.ui.model
 
 enum class Color {
     RED, ORANGE, YELLOW, GREEN, BLUE
