+++ app/src/main/java/woowacourse/paint/presentation/ui/uiState/MainUiState.kt
@@ -1,8 +1,8 @@
 package woowacourse.paint.presentation.ui.uiState
 
-import woowacourse.paint.data.model.Brush
-import woowacourse.paint.data.model.Color
-import woowacourse.paint.data.model.SettingMode
+import woowacourse.paint.presentation.ui.model.Brush
+import woowacourse.paint.presentation.ui.model.Color
+import woowacourse.paint.presentation.ui.model.SettingMode
 
 data class MainUiState(
     val settingMode: SettingMode,
